test: :white_check_mark: Adiciona testes para view de alteracao de nome e renomea arquivo
